# docs/contracts/http/schemaapi.openapi.yaml
openapi: 3.0.3
info:
  title: SchemaAPI
  version: 0.5.0
  description: Document Intelligence API with ingestion, hybrid retrieval, cited RAG, governance, agents, analysis reports and observability.
servers:
  - url: http://localhost:8081
paths:
...
  /analysis/reports:
    post:
      summary: Create an analysis report from selected documents and questions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                document_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                rag_queries:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Analysis report
    get:
      summary: List analysis reports
      responses:
        "200":
          description: Analysis report list
  /analysis/reports/{id}:
    get:
      summary: Get an analysis report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Analysis report
        "404":
          description: Analysis report not found
  /analysis/reports/{id}/export:
    get:
      summary: Export an analysis report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Exported report
        "404":
          description: Analysis report not found
...